+++ tests/unit/query_common.hpp
@@ -597,3 +597,7 @@ auto GetMerge(AstTreeStorage &storage, Pattern *pattern, OnMatch on_match,
 #define ALL(variable, list, where)                                        \
   storage.Create<query::All>(storage.Create<query::Identifier>(variable), \
                              list, where)
+#define REDUCE(accumulator, initializer, variable, list, expr)     \
+  storage.Create<query::Reduce>(                                   \
+      storage.Create<query::Identifier>(accumulator), initializer, \
+      storage.Create<query::Identifier>(variable), list, expr)